backend/routingapp/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
PyJWT==2.8.0
python-decouple==3.8
Requests==2.32.3
psycopg2-binary==2.9.9
pika==1.2.0
celery==5.4.0
django-celery-results==2.5.1
polyline==2.0.2
